Fix the cref type of lemmas
The lemmas share the same counter with the theorems.
When cref is used to refer to a lemma,
the type is wrongly displayed as a theorem.
This is because package cleveref has to be loaded after newtheorem commands.

Move the loading of cleveref package as well as other common packages to the format file.